+++ HISTORY.rst
@@ -7,9 +7,9 @@ TODO
 * Add more meaningful unittests.
 * Export heavy definition logic (such as field definition) from the jinja template to python filters.
 
-In Progress
------------
-* Add more meaningful unittests.
+Read for Release
+----------------
+* Added more meaningful unittests.
 
 
 0.2.0 (2019-01-14)